Export of internal changes
--
310930395 by tswadell <tswadell@google.com>:

BEGIN_PUBLIC
Add alwayslink = 1 to ANTLR Bazel target to support OSX builds.

The flag is not necessary for Linux builds, but seems to cause
no harm when added and makes it so that local builds on OSX work.
END_PUBLIC

This is a no-op for google3 builds.
--
310915802 by tswadell <tswadell@google.com>:

BEGIN_PUBLIC
Upgrade the ANTLR runtime to 4.7.2
END_PUBLIC

This change does not affect the google3 builds which use 4.7.1.
Given the extent of 4.7.1's use the google3 upgrade would be a
considerable effort that is not urgent.
--
310910752 by CEL Dev Team <cel-dev@google.com>:

BEGIN_PUBLIC
Add missing override for CEL parser visitor visitConstantLiteral()

The default implementation is incompatible with Antlr 4.7.2. An upgrade is
desirable to benefit from thread-safety improvements.
END_PUBLIC
--
309264125 by CEL Dev Team <cel-dev@google.com>:

BEGIN_PUBLIC
Internal change
END_PUBLIC

- Make Parser visible to CAA
--
309098060 by CEL Dev Team <cel-dev@google.com>:

LSC: Allow targets to pass with `-fsanitize=null` under `--config=asan`.

This is a manually-generated change. Ensure that behavior and style are
preserved by carefully inspecting the diff.

This CL fixes one of the many targets in TAP that test with ASan but fail
with `-fsanitize=null`. We would like to enable the Null Pointer Sanitizer
in builds that use Address Sanitizer globally, and this change gets us one
step closer to that goal.

See http://go/sanitize-null-cleanup-lsc for details.
--
309036708 by CEL Dev Team <cel-dev@google.com>:

BEGIN_PUBLIC
Moves set_util from cel/cpp/eval/eval to cel/cpp/eval/public
END_PUBLIC
--
308324817 by CEL Dev Team <cel-dev@google.com>:

BEGIN_PUBLIC
Update EnrichSourceInfo to contain a map of node_id with positions
END_PUBLIC
--
308317802 by CEL Dev Team <cel-dev@google.com>:

LSC: Fix incorrect string escapes in Copybara config files

For example, the regex "\w" is invalid. It should be either: r"\w" or "\\w"

This CL does not change behavior: Copybara currently treats invalid
escapes (like "\w") equivalently to the corrected versions (like "\\w").

Copybara is going to enable strict Starlark validation. As part
of this change we are going to enable restrictStringEscapes. This flag
prevents users from using incorrect escapes.

This was done by running buildifier copy.bara.sky

More information: go/copybara-strict-string-escapes

BEGIN_PUBLIC
Internal change
END_PUBLIC

Tested:
    TAP --sample ran all affected tests and none failed
    http://test/OCL:308094157:BASE:308146641:1587694686416:2641e69f
--
308129984 by CEL Dev Team <cel-dev@google.com>:

BEGIN_PUBLIC
Extends the CelValueLessThan Visitor implementation to handle less than, deep equality and greater than in a single pass.
Adds functions to check for deep equality and greater than of two CelValues using the extended Visitor.
END_PUBLIC
--
307113050 by CEL Dev Team <cel-dev@google.com>:

LSC: Mark static const class/struct members as constexpr. This change fixes declarations that have initial values but are technically not definitions by marking them constexpr (which counts as a definition). This enables, among other things, the modified constants to be passed into functions and function templates that accept arguments by reference. Without this change, such functions would cause linker errors.

This change was approved as part of go/static-const-lsc.

Tested:
    TAP train for global presubmit queue
    http://test/OCL:306269815:BASE:306377582:1586865021582:ea5fa059
--
307074803 by tswadell <tswadell@google.com>:

Delete the legacy conformance directory under google/api/expr/conformance

BEGIN_PUBLIC
Build visibility change
END_PUBLIC
--
306984679 by kuat <kuat@google.com>:

BEGIN_PUBLIC
Internal change.
END_PUBLIC

Remove //base dependency.
Fix gcc compiler warnings.
--
306926648 by CEL Dev Team <cel-dev@google.com>:

BEGIN_PUBLIC
Bugfix: return reference of EnrichedSourceInfo in getter.
END_PUBLIC

- Make Parser visible to caa troubleshooter
--
306564608 by CEL Dev Team <cel-dev@google.com>:

BEGIN_PUBLIC
Add EnrichedParse function to return enriched response with token end indexes.
END_PUBLIC
--
306523921 by kuat <kuat@google.com>:

BEGIN_PUBLIC
refactor: provide default implementations for BaseActivation.
refactor: return FieldMask by reference to avoid a protobuf copy.
END_PUBLIC
--
306335623 by CEL Dev Team <cel-dev@google.com>:

BEGIN_PUBLIC
Move explainer transform utility class to public path
END_PUBLIC

PiperOrigin-RevId: 310930395

+++ eval/eval/function_step.cc
@@ -47,7 +47,7 @@ bool ShouldAcceptOverload(const CelFunction* function,
   if (function == nullptr) {
     return false;
   }
-  for (int i = 0; i < arguments.size(); i++) {
+  for (size_t i = 0; i < arguments.size(); i++) {
     if (arguments[i].IsUnknownSet() || arguments[i].IsError()) {
       return IsNonStrict(function->descriptor().name());
     }
@@ -65,7 +65,7 @@ std::vector<CelValue> CheckForPartialUnknowns(
     absl::Span<const AttributeTrail> attrs) {
   std::vector<CelValue> result;
   result.reserve(args.size());
-  for (int i = 0; i < args.size(); i++) {
+  for (size_t i = 0; i < args.size(); i++) {
     auto attr_set = frame->unknowns_utility().CheckForUnknowns(
         attrs.subspan(i, 1), /*use_partial=*/true);
     if (!attr_set.attributes().empty()) {
